+++ src/dex/base-dex.ts
@@ -1,17 +1,12 @@
 import { LiquidityPool } from './models/liquidity-pool';
 import { BaseDataProvider } from '@providers/data/base-data-provider';
 import { Asset, Token } from './models/asset';
-import { AssetBalance, DatumParameters, PayToAddress, SwapFee, Transaction, UTxO } from '@app/types';
+import { AssetBalance, DatumParameters, PayToAddress, SwapFee, UTxO } from '@app/types';
 import { DatumParameterKey } from '@app/constants';
 import { tokensMatch } from '@app/utils';
 import { BaseApi } from '@dex/api/base-api';
 
 export abstract class BaseDex {
-
-    /**
-     * Unique name for the DEX.
-     */
-    public abstract readonly name: string;
 
     /**
      * API connection for the DEX.