+++ src/dex/models/liquidity-pool.ts
@@ -29,10 +29,10 @@ export class LiquidityPool {
     }
 
     get uuid(): string {
-        return `${this.dex}.${this.name}.${this.identifier}`;
+        return `${this.dex}.${this.pair}.${this.identifier}`;
     }
 
-    get name(): string {
+    get pair(): string {
         const assetAName: string = this.assetA === 'lovelace' ? 'ADA' : this.assetA.assetName;
         const assetBName: string = this.assetB === 'lovelace' ? 'ADA' : this.assetB.assetName;